iio: adc: ad9361: opaque-ify some drv structs and split state struct
The much-awaited commit that hides aways platform data away from
`ad9361_conv.c` and splits the state information into a separate struct.
The 2 are connected a bit; splitting the struct does not make without
making it opaque; even though possible, it's a bit error prone, because the
compiler can help quite a bit to determine what's exposed [and what isn't].
